feat: always use float64 instead of float32

+++ integration/pkg/api/osrmv1/route_response.go
@@ -11,19 +11,19 @@ type RouteResponse struct {
 
 // Route represents a route through (potentially multiple) waypoints.
 type Route struct {
-	Distance   float32     `json:"distance"`
-	Duration   float32     `json:"duration"`
+	Distance   float64     `json:"distance"`
+	Duration   float64     `json:"duration"`
 	Geometry   string      `json:"geometry"`
-	Weight     float32     `json:"weight"`
+	Weight     float64     `json:"weight"`
 	WeightName string      `json:"weight_name"`
 	Legs       []*RouteLeg `json:"legs,omitempty"`
 }
 
 // RouteLeg represents a route between two waypoints.
 type RouteLeg struct {
-	Distance   float32      `json:"distance"`
-	Duration   float32      `json:"duration"`
-	Weight     float32      `json:"weight"`
+	Distance   float64      `json:"distance"`
+	Duration   float64      `json:"duration"`
+	Weight     float64      `json:"weight"`
 	Summary    string       `json:"summary"`
 	Steps      []*RouteStep `json:"steps,omitempty"`
 	Annotation Annotation   `json:"annotation"`
@@ -31,10 +31,10 @@ type RouteLeg struct {
 
 // RouteStep A step consists of a maneuver such as a turn or merge, followed by a distance of travel along a single way to the subsequent step.
 type RouteStep struct {
-	Distance float32 `json:"distance"`
-	Duration float32 `json:"duration"`
+	Distance float64 `json:"distance"`
+	Duration float64 `json:"duration"`
 	Geometry string  `json:"geometry"`
-	Weight   float32 `json:"weight"`
+	Weight   float64 `json:"weight"`
 	Name     string  `json:"name"`
 	//TODO: others
 }
@@ -42,19 +42,19 @@ type RouteStep struct {
 // Waypoint object used to describe waypoint on a route.
 type Waypoint struct {
 	Name     string     `json:"name"`
-	Location [2]float32 `json:"location,omitempty"` // [longitude, latitude]
-	Distance float32    `json:"distance"`
+	Location [2]float64 `json:"location,omitempty"` // [longitude, latitude]
+	Distance float64    `json:"distance"`
 	Hint     string     `json:"hint"`
 }
 
 // Annotation of the whole route leg with fine-grained information about each segment or node id.
 type Annotation struct {
-	Distance    []float32 `json:"distance,omitempty"`
-	Duration    []float32 `json:"duration,omitempty"`
+	Distance    []float64 `json:"distance,omitempty"`
+	Duration    []float64 `json:"duration,omitempty"`
 	DataSources []int     `json:"datasources,omitempty"`
 	Nodes       []int64   `json:"nodes,omitempty"`
-	Weight      []float32 `json:"weight,omitempty"`
-	Speed       []float32 `json:"speed,omitempty"`
+	Weight      []float64 `json:"weight,omitempty"`
+	Speed       []float64 `json:"speed,omitempty"`
 	Metadata    Metadata  `json:"metadata,omitempty"`
 }
